Refactor paginator/main js to be more compartmentalized
